+++ src/extension/endpoint/http-endpoint/http-endpoint.ts
@@ -1,8 +1,9 @@
 import axios, { AxiosInstance, AxiosRequestConfig } from 'axios';
 
-import { getSPARQLQueryKind, getAcceptHeader } from '../sparql-utils';
-import { Endpoint } from '../endpoint';
+import { getAcceptHeader } from '../sparql-utils';
+import { Endpoint, SimpleHttpResponse } from '../endpoint';
 import { SparqlQuery } from '../model/sparql-query';
+import { MimeType } from '../../enum/mime-type';
 
 /**
  * Represents an HTTP SPARQL endpoint.
@@ -38,7 +39,7 @@ export class HttpEndpoint extends Endpoint {
    * @param sparqlQuery - The SPARQL query to execute.
    * @param execution - The execution object.
    */
-  public async query(sparqlQuery: SparqlQuery, execution?: any) {
+  public async query(sparqlQuery: SparqlQuery, execution?: any): Promise<SimpleHttpResponse> {
     const params = new URLSearchParams();
     params.append("query", sparqlQuery.queryString);
 
@@ -61,7 +62,13 @@ export class HttpEndpoint extends Endpoint {
       });
     }
     const response = await this.http.post('', params, options);
-    return response;
+    const mimeType = response.headers['content-type'];
+
+    const httpResponse: SimpleHttpResponse = {
+      headers: { "content-type": mimeType },
+      data: typeof response.data === 'string' ? response.data : JSON.stringify(response.data)
+    };
+    return httpResponse;
   }
 
 }