+++ src/extension/enum/mime-type.ts
@@ -1,0 +1,14 @@
+export enum MimeType {
+    turtle = 'text/turtle',
+    ntriples = 'application/n-triples',
+    rdfxml = 'application/rdf+xml',
+    trig = 'application/trig',
+    nquads = 'application/n-quads',
+    json = 'application/json',
+    sparqlResultsJson = 'application/sparql-results+json',
+    sparqlResultsXml = 'application/sparql-results+xml',
+    sparqlResultsCsv = 'text/csv',
+    sparqlResultsTsv = 'text/tab-separated-values',
+    plainText = 'text/plain',
+    any = '*/*',
+}