+++ src/extension/notebook/sparql-notebook-controller.ts
@@ -6,6 +6,8 @@ import { notebookEndpoint } from '../endpoint/endpoint';
 import { SparqlNotebookCell } from './sparql-notebook-cell';
 import { shrink } from '@zazuko/prefixes';
 import { SparqlQuery } from '../endpoint/model/sparql-query';
+import { MimeType } from '../enum/mime-type';
+import { MIMEParams, MIMEType } from 'util';
 
 export class SparqlNotebookController {
   readonly controllerId = `${extensionId}-controller-id`;
@@ -109,8 +111,9 @@ export class SparqlNotebookController {
     const data = queryResult.data;
     let isSuccess = true;
 
-    if (contentType === "application/sparql-results+json") {
-      if (data.hasOwnProperty("boolean")) {
+    if (contentType === MimeType.sparqlResultsJson) {
+      const parsedData = JSON.parse(data);
+      if (parsedData.hasOwnProperty("boolean")) {
         // sparql ask
         execution.replaceOutput([this._writeAskResult(data)]);
         execution.end(isSuccess, Date.now());
@@ -123,16 +126,17 @@ export class SparqlNotebookController {
       return;
     }
 
-    if (contentType === "text/turtle") {
+    if (contentType === MimeType.turtle) {
       // sparql construct
-      execution.replaceOutput([await this._writeTurtleResult(data, sparqlCell.getPrefixMap())]);
+      execution.replaceOutput([await this._writeTurtleResult(data)]);
       execution.end(isSuccess, Date.now());
       return;
     }
 
-    if (contentType === "application/json") {
+    if (contentType === MimeType.json) {
       // stardog is returning and error as json
-      execution.replaceOutput([this._writeError(data.message)]);
+      const errorObject = JSON.parse(data);
+      execution.replaceOutput([this._writeError(errorObject.message ?? 'An error occurred')]);
       isSuccess = false;
       execution.end(isSuccess, Date.now());
       return;
@@ -146,12 +150,12 @@ export class SparqlNotebookController {
     return;
   }
 
-  private async _writeTurtleResult(resultTTL: string, prefix: PrefixMap): Promise<NotebookCellOutput> {
+  private async _writeTurtleResult(resultTTL: string): Promise<NotebookCellOutput> {
 
     // this is writing markdown to the cell containing a turtle code block
     return new NotebookCellOutput([
 
-      NotebookCellOutputItem.text(resultTTL, "text/plain"),
+      NotebookCellOutputItem.text(resultTTL, MimeType.plainText),
       NotebookCellOutputItem.text(
         `\`\`\`turtle\n${resultTTL}\n\`\`\``,
         "text/markdown"
@@ -159,23 +163,23 @@ export class SparqlNotebookController {
     ]);
   }
 
-  private _writeAskResult(resultJson: any): NotebookCellOutput {
+  private _writeAskResult(resultJson: string): NotebookCellOutput {
     const outputItem = new NotebookCellOutput([
-      this._writeJson(JSON.stringify(resultJson, null, "   ")),
+      this._writeJson(resultJson),
       NotebookCellOutputItem.json(
-        resultJson,
-        "application/sparql-results+json"
+        JSON.parse(resultJson),
+        MimeType.sparqlResultsJson
       ),
     ]);
     return outputItem;
   }
 
-  private _writeSparqlJsonResult(resultJson: any, prefixMap: PrefixMap = {}): NotebookCellOutput {
+  private _writeSparqlJsonResult(resultJson: string, prefixMap: PrefixMap = {}): NotebookCellOutput {
     const outputItem = new NotebookCellOutput([
-      this._writeJson(JSON.stringify(resultJson, null, "   ")),
+      this._writeJson(resultJson),
       NotebookCellOutputItem.json(
-        resultJson,
-        "application/sparql-results+json"
+        JSON.parse(resultJson),
+        MimeType.sparqlResultsJson
       ),
       this._writeDataTableRendererCompatibleJson(resultJson, prefixMap)
     ]);
@@ -183,14 +187,14 @@ export class SparqlNotebookController {
     return outputItem;
   }
 
-  private _writeJson(jsonResult: any): NotebookCellOutputItem {
+  private _writeJson(jsonResult: string): NotebookCellOutputItem {
     return NotebookCellOutputItem.text(
       `\`\`\`json\n${jsonResult}\n\`\`\``,
       "text/markdown"
     );
   }
 
-  private _writeError(message: any): NotebookCellOutput {
+  private _writeError(message: string): NotebookCellOutput {
     return new NotebookCellOutput([
       NotebookCellOutputItem.error({
         name: "SPARQL error",
@@ -199,7 +203,8 @@ export class SparqlNotebookController {
     ]);
   }
 
-  private _writeDataTableRendererCompatibleJson(resultJson: any, prefixMap: PrefixMap = {}) {
+  private _writeDataTableRendererCompatibleJson(resultJsonString: string, prefixMap: PrefixMap = {}) {
+    const resultJson = JSON.parse(resultJsonString);
     const dtJonBindings: { [k: string]: any }[] = resultJson.results.bindings;
 
     const dtJson = dtJonBindings.map(item => {
@@ -220,7 +225,7 @@ export class SparqlNotebookController {
     });
 
     const jsonStringified = JSON.stringify(dtJson, null, "   ");
-    return NotebookCellOutputItem.text(jsonStringified, "application/json");
+    return NotebookCellOutputItem.text(jsonStringified, MimeType.json);
   }
 
   dispose() {
@@ -249,7 +254,6 @@ export class SparqlNotebookController {
         const activeFileDir = Uri.parse(cell.document.uri.toString(true)).with({ path: cell.document.uri.path.replace(/\/[^\/]*$/, '') });
         fileUri = activeFileDir.with({ path: `${activeFileDir.path}/${filePath}` });
       }
-      console.log('fileUri', fileUri.fsPath);
       const fileEndpoint = new FileEndpoint();
       await fileEndpoint.addFile(fileUri);
       return fileEndpoint;