+++ src/extension/sparql-store/sparql-store.ts
@@ -1,4 +1,4 @@
-import { Store, defaultGraph } from 'oxigraph';
+import { Quad, Store, defaultGraph } from 'oxigraph';
 import { SPARQLQueryKind } from '../endpoint/enum/sparql-query-kind';
 import { SparqlQuery } from '../endpoint/model/sparql-query';
 
@@ -9,6 +9,11 @@ export enum RdfMimeType {
     trig = 'application/trig',
     nQuads = 'application/n-quads',
 }
+
+/**
+ * This is the local SPARQL endpoint that is used to execute SPARQL queries on the local RDF store.
+ * It is based on the Oxigraph.
+ */
 export class SparqlStore {
     // the oxigraph store
     readonly #store: Store;
@@ -18,59 +23,107 @@ export class SparqlStore {
     }
 
     /**
-     * Can be a SELECT, ASK, CONSTRUCT or DESCRIBE query
+     * Execute a SPARQL CONSTRUCT query and return the result as a turtle string.
      * 
-     * @param query SPARQL query
-     * @returns 
+     * @param query SPARQL query of type CONSTRUCT
+     * @returns the turtle string
+     * @throws Error if the query is not a CONSTRUCT query
      */
-    public async query(query: SparqlQuery): Promise<any> {
-        // Executes a SPARQL 1.1 Query. 
-        // For SELECT queries the return type is an array of Map which keys are the bound variables and values are the values the result is bound to. 
-        // For CONSTRUCT and DESCRIBE queries the return type is an array of Quad. 
-        // For ASK queries the return type is a boolean.
-
-        const queryKind = query.kind;
-        switch (queryKind) {
-            case SPARQLQueryKind.ask:
-            case SPARQLQueryKind.select:
-                const res = await this._select(query);
-                const fakeHttpResult = {
-                    headers: { "content-type": "application/sparql-results+json" },
-                    data: JSON.parse(res)
-                };
-                return fakeHttpResult;
-            case SPARQLQueryKind.construct:
-            case SPARQLQueryKind.describe:
-                return this._construct(query);
-            default:
-                return await this.#store.query(query.queryString, {});
+    construct(query: SparqlQuery): string | never {
+        if (query.kind !== SPARQLQueryKind.construct) {
+            throw new Error('Query is not a CONSTRUCT query');
         }
 
-    }
+        const queryResult = this.#store.query(query.queryString, {
+            results_format: RdfMimeType.turtle,
+        }) as string;
 
-    private async _construct(query: SparqlQuery): Promise<any> {
-        const ttl = (new Store(this.#store.query(query.queryString, {}))).dump(RdfMimeType.turtle, defaultGraph());
-        const fakeHttpResult = {
-            // eslint-disable-next-line @typescript-eslint/naming-convention
-            headers: { "content-type": RdfMimeType.turtle },
-            data: ttl
-        };
-        return fakeHttpResult;
-    }
-
-    private async _select(query: SparqlQuery): Promise<string> {
-        const result = await this.#store.query(query.queryString, {
-            results_format: "json",
-        });
-        return result;
-    };
-
-
-    public async update(query: SparqlQuery): Promise<any> {
-        return await this.#store.update(query.queryString, {});
+        return queryResult;
     }
 
     /**
+     * Execute a SPARQL CONSTRUCT query and return the result as an array of quads.
+     * 
+     * @param query SPARQL query of type CONSTRUCT
+     * @returns the array of quads
+     * @throws Error if the query is not a CONSTRUCT query
+     */
+    constructQuads(query: SparqlQuery): Quad[] | never {
+        if (query.kind !== SPARQLQueryKind.construct) {
+            throw new Error('Query is not a CONSTRUCT query');
+        }
+
+        const queryResult = this.#store.query(query.queryString) as Quad[];
+
+        return queryResult;
+    }
+
+    /**
+     * Execute a SPARQL SELECT query and return the result as a JSON string.
+     * 
+     * @param query SPARQL query of type SELECT
+     * @returns the JSON string
+     * @throws Error if the query is not a SELECT query
+     */
+    select(query: SparqlQuery): string | never {
+        if (query.kind !== SPARQLQueryKind.select) {
+            throw new Error('Query is not a SELECT query');
+        }
+
+        const queryResult = this.#store.query(query.queryString, {
+            results_format: 'json',
+        }) as string;
+
+        return queryResult;
+    }
+
+    /**
+     * Execute a SPARQL ASK query
+     * 
+     * @param query SPARQL query of type ASK
+     * @returns the JSON string
+     */
+    ask(query: SparqlQuery): string | never {
+        if (query.kind !== SPARQLQueryKind.ask) {
+            throw new Error('Query is not an ASK query');
+        }
+        const result = this.#store.query(query.queryString, {
+            results_format: "json",
+        }) as string;
+        return result;
+    }
+
+    /**
+     * Execute a SPARQL DESCRIBE query and return the result as a turtle string.
+     * 
+     * @param query SPARQL query of type DESCRIBE
+     * @returns the turtle string
+     */
+    describe(query: SparqlQuery): string | never {
+        if (query.kind !== SPARQLQueryKind.describe) {
+            throw new Error('Query is not a DESCRIBE query');
+        }
+
+        const queryResult = this.#store.query(query.queryString, {
+            results_format: RdfMimeType.turtle,
+        }) as string;
+
+        return queryResult;
+    }
+
+    describeQuads(query: SparqlQuery): Quad[] | never {
+        if (query.kind !== SPARQLQueryKind.describe) {
+            throw new Error('Query is not a DESCRIBE query');
+        }
+
+        const queryResult = this.#store.query(query.queryString) as Quad[];
+
+        return queryResult;
+    }
+
+
+    /**
+     * Load an RDF string into the store.
      * 
      * @param rdfString Load RDF data into the store
      * @param mimeType Supported MIME types are: application/n-triples, text/turtle, application/rdf+xml, application/n-quads
@@ -82,22 +135,22 @@ export class SparqlStore {
             to_graph_name: defaultGraph()
         };
 
-        return this.#store.load(rdfString, options, null, null);
+        this.#store.load(rdfString, options);
     }
 
     /**
+     * Dump the store into an RDF string.
      * 
      * @param mimeType Supported MIME types are: application/n-triples, text/turtle, application/rdf+xml, application/n-quads
      * @returns Serialized RDF data
      */
-    public async dump(mimeType: RdfMimeType): Promise<string> {
+    public dump(mimeType: RdfMimeType): string {
         const options = {
             format: mimeType,
             to_graph_name: defaultGraph()
         };
-        return await this.#store.dump(options, null);
+        return this.#store.dump(options);
     }
-
 
     /**
      * Get the size of the store
@@ -108,3 +161,11 @@ export class SparqlStore {
     }
 
 }
+
+
+/**
+     const fakeHttpResult = {
+                    headers: { "content-type": "application/sparql-results+json" },
+                    data: JSON.parse(res)
+                };
+ */